Use Decimal field to story Item amount

+++ items/models.py
@@ -9,7 +9,7 @@ class Item(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     create_date = models.DateTimeField(auto_now_add=True)
     update_date = models.DateTimeField(auto_now=True)
-    amount = models.BigIntegerField()
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
 
 
 class TransactionEnum(Enum):
